+++ src/test/java/old/BaseTest.java
@@ -1,0 +1,25 @@
+package old;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+
+import java.util.concurrent.TimeUnit;
+
+public class BaseTest {
+
+    public WebDriver driver;
+    @BeforeEach
+    public void setUp() {
+        System.setProperty("webdriver.chrome.driver", "/Users/tatiana/mySecondProjectAutomation/src/main/resources/drivers/chromedriver");
+        driver = new ChromeDriver();
+        driver.manage().window().maximize();
+        driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
+        driver.get("https://github.com/login");
+    }
+    @AfterEach
+    public void tearDown() {
+        driver.quit();
+    }
+}
